requirements.txt
aiohttp
aiogram
apscheduler
python-dotenv
